+++ seedfarmer/commands/_deployment_commands.py
@@ -374,15 +374,14 @@ def _deploy_validated_deployment(
 
 
 def prime_target_accounts(deployment_manifest: DeploymentManifest) -> None:
-    # TODO: Investigate whether we need to validate the requested mappings against previously deployed mappings
-
     _logger.info("Priming Accounts")
     with concurrent.futures.ThreadPoolExecutor(max_workers=len(deployment_manifest.target_accounts_regions)) as workers:
 
-        def _prime_accounts(args: Dict[str, Any]) -> None:
+        def _prime_accounts(args: Dict[str, Any]) -> List[Any]:
             _logger.info("Priming Acccount %s in %s", args["account_id"], args["region"])
-            commands.deploy_seedkit(**args)
+            seedkit_stack_outputs = commands.deploy_seedkit(**args)
             commands.deploy_managed_policy_stack(deployment_manifest=deployment_manifest, **args)
+            return [args["account_id"], args["region"], seedkit_stack_outputs]
 
         params = []
         for target_account_region in deployment_manifest.target_accounts_regions:
@@ -401,7 +400,11 @@ def prime_target_accounts(deployment_manifest: DeploymentManifest) -> None:
 
             params.append(param_d)
 
-        _ = list(workers.map(_prime_accounts, params))
+        output_seedkit = list(workers.map(_prime_accounts, params))
+        # add these to the region mappings for reference
+        for out_s in output_seedkit:
+            deployment_manifest.populate_seedkit_metadata(account_id=out_s[0], region=out_s[1], seedkit_dict=out_s[2])
+        _logger.debug(deployment_manifest.model_dump())
 
 
 def tear_down_target_accounts(deployment_manifest: DeploymentManifest, retain_seedkit: bool = False) -> None:
@@ -551,7 +554,7 @@ def deploy_deployment(
 
         By default False
     """
-    deployment_manifest_wip = deployment_manifest.copy()
+    deployment_manifest_wip = deployment_manifest.model_copy()
     deployment_name = cast(str, deployment_manifest_wip.name)
     _logger.debug("Setting up deployment for %s", deployment_name)
 
@@ -604,7 +607,9 @@ def deploy_deployment(
                 deployment_manifest=deployment_manifest_wip, module=module, group_name=group.name
             )
 
-            module.manifest_md5 = hashlib.md5(json.dumps(module.dict(), sort_keys=True).encode("utf-8")).hexdigest()
+            module.manifest_md5 = hashlib.md5(
+                json.dumps(module.model_dump(), sort_keys=True).encode("utf-8")
+            ).hexdigest()
             module.deployspec_md5 = hashlib.md5(open(deployspec_path, "rb").read()).hexdigest()
 
             _build_module = du.need_to_build(
@@ -704,7 +709,7 @@ def apply(
     manifest_path = os.path.join(config.OPS_ROOT, deployment_manifest_path)
     with open(manifest_path) as manifest_file:
         deployment_manifest = DeploymentManifest(**yaml.safe_load(manifest_file))
-    _logger.debug(deployment_manifest.dict())
+    _logger.debug(deployment_manifest.model_dump())
 
     # Initialize the SessionManager for the entire project
     session_manager = SessionManager().get_or_create(
@@ -720,7 +725,9 @@ def apply(
     deployment_manifest._partition = partition
     if not dryrun:
         write_deployment_manifest(
-            cast(str, deployment_manifest.name), deployment_manifest.dict(), session=session_manager.toolchain_session
+            cast(str, deployment_manifest.name),
+            deployment_manifest.model_dump(),
+            session=session_manager.toolchain_session,
         )
 
     for module_group in deployment_manifest.groups: