+++ seedfarmer/models/_base.py
@@ -15,7 +15,7 @@
 from typing import Optional, cast
 
 import humps
-from pydantic import BaseModel
+from pydantic import BaseModel, ConfigDict
 
 
 def to_camel(string: str) -> str:
@@ -23,10 +23,9 @@ def to_camel(string: str) -> str:
 
 
 class CamelModel(BaseModel):
-    class Config:
-        alias_generator = to_camel
-        allow_population_by_field_name = True
-        underscore_attrs_are_private = True
+    # TODO[pydantic]: The following keys were removed: `underscore_attrs_are_private`.
+    # Check https://docs.pydantic.dev/dev-v2/migration/#changes-to-config for more information.
+    model_config = ConfigDict(alias_generator=to_camel, populate_by_name=True)
 
 
 class ModuleRef(CamelModel):