+++ CHANGELOG.md
@@ -7,8 +7,11 @@ This project adheres to [Semantic Versioning](http://semver.org/) and [Keep a Ch
 ## Unreleased
 
 ### New
+- adding AWS curated codebuild iamge override with opinionated runtimes
 
 ### Changes
+- updating pydantic support from 1.X.X to 2.5.3
+- adding seedfarmer verions check support with `seedfarmer.yaml`
 
 ### Fixes
 - update `manifests/examples/` to point to an updated release branch